requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.6
distro==1.9.0
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.5.0
gradio==4.32.0
gradio-client==0.17.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-nvidia-ai-endpoints==0.1.0
langchain-text-splitters==0.2.0
langchainhub==0.1.16
langsmith==0.1.63
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.3
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.1
pydantic-core==2.18.2
pydub==0.25.1
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.32.2
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
spotipy==2.23.0
SQLAlchemy==2.0.30
starlette==0.37.2
tenacity==8.3.0
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
typer==0.12.3
types-requests==2.32.0.20240523
typing-extensions==4.12.0
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
watchfiles==0.22.0
websockets==11.0.3
yarl==1.9.4
zipp==3.19.0
